// src/modules/data-processing/last-30-days.interface.ts
export interface Last30Days {
    period_name: string;
    total_birds_affected: number; // Total birds affected
    total_flocks_affected: number; // Total flocks affected
    total_backyard_flocks_affected: number; // Backyard flocks affected
    total_commercial_flocks_affected: number; // Commercial flocks affected
}